refactor: move yaml settings source to pydantic module
