+++ kpops/cli/main.py
@@ -11,7 +11,6 @@ import typer
 from kpops import __version__
 from kpops.cli.custom_formatter import CustomFormatter
 from kpops.cli.registry import Registry
-from kpops.cli.settings_sources import YamlConfigSettingsSource
 from kpops.component_handlers import ComponentHandlers
 from kpops.component_handlers.kafka_connect.kafka_connect_handler import (
     KafkaConnectHandler,
@@ -22,6 +21,7 @@ from kpops.component_handlers.topic.proxy_wrapper import ProxyWrapper
 from kpops.config import KpopsConfig
 from kpops.pipeline_generator.pipeline import Pipeline
 from kpops.utils.gen_schema import SchemaScope, gen_config_schema, gen_pipeline_schema
+from kpops.utils.pydantic import YamlConfigSettingsSource
 
 if TYPE_CHECKING:
     from collections.abc import Iterator
